refactor: change results and storys to utils folder
