+++ src/app/game/page.tsx
@@ -1,9 +1,9 @@
 import GamePage from "@/components/Game/GamePage";
 import MultipleChoice from "@/components/Game/MultipleChoice";
 import ResultGame from "@/components/Game/ResultGame";
-import Results from "@/components/Game/Results";
+import Results from "@/utils/Results";
 import StoryCard from "@/components/Game/StoryCard";
-import Storys from "@/components/Game/Storys";
+import Storys from "@/utils/Storys";
 import GameLayout from "@/layouts/GameLayout";
 import { useEffect, useState } from "react";
